Django-Middleware-0x03/requirements.txt
absl-py==2.3.1
aiosqlite==0.21.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.11
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.10
exceptiongroup==1.3.0
flatbuffers==25.2.10
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jax==0.4.30
jaxlib==0.4.30
joblib==1.2.0
kiwisolver==1.4.7
kombu==5.5.4
matplotlib==3.9.4
mediapipe==0.10.14
ml_dtypes==0.5.3
MouseInfo==0.1.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.0.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
parameterized==0.9.0
pillow==10.4.0
pluggy==1.6.0
profanity-check==1.0.3
prompt_toolkit==3.0.52
protobuf==4.25.8
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.23
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
PyMySQL==1.1.2
pynput==1.8.1
pyparsing==3.2.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.5
scikit-learn==0.24.2
scipy==1.13.1
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.1
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zipp==3.23.0
